Add hover-hide-delay functionality to pop-ups
With this CL, the `hover-pop-up-hide-delay` CSS property can be used
to make a pop-up hide after a specified delay, if it is no longer
hovered. This is related, but independent from, the hover-pop-up-delay
functionality, which can be used on *another element* to trigger/show
a pop-up after a delay. The `hover-pop-up-hide-delay` can be used on
any pop-up type, including popup=manual, and it will make the pop-up
get automatically hidden if it (or its anchors or invokers) is not
hovered for the specified period of time.

A typical use case is to build a tooltip, fully in HTML/CSS:

  <input hoverpopup=tooltip placeholder="Hover me">
  <div popup=hint role=tooltip id=tooltip>Additional info</div>

  <style>
    input {
      hover-pop-up-delay: 0.5s;
    }
    [popup] {
      hover-pop-up-hide-delay: 1s;
      opacity:0;
      transition: opacity 0.5s;
    }
    [popup]:top-layer {
      opacity:1;
    }
  </style>

Bug: 1307772
Change-Id: I12b36cf0e612b541c054a9a41c6d91f58c55ec72
Cq-Do-Not-Cancel-Tryjobs: true