Fix offsetTop/offsetLeft for new offsetParent behavior
While going up the flat tree to find an offsetParent,
HTMLElement.offsetParent now skips elements which are closed shadow
hidden from the element behind the OffsetParentNewSpecBehavior flag.

However, there is additional correction needed in offsetTop and
offsetLeft. They are supposed to be corresponding to the offsetParent,
but I didn't implement this correction when I first implemented
OffsetParentNewSpecBehavior.

This patch makes offsetTop and offsetLeft add up all of the intermediate
values skipped by the new offsetParent logic. It is similar to the
WebKit implementation: https://trac.webkit.org/changeset/239313/webkit

This patch also re-enables the flag now that all known issues are fixed.

Bug: 1330819, 920069
Change-Id: I6183909a4dfa763f2f3e178edd0541866cd5bbf6