Add RemoteContextHelper a convenience wrapper for remote contexts.
This provides a way to easily create them, runs scripts in them and correctly navigate them.

It also adds 2 tests that use it to test unload handlers in the main frame. I could not find an existing test (the ones alongside all test unload in an iframe).

Some more tests using this are in https://crrev.com/c/3750822.

Bug: 1348500
Change-Id: I5b0bd107283bf54c75b4eed9b276728581d7d557